backend/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
SQLAlchemy==1.4.46
alembic==1.10.2
pydantic==1.10.7
PyJWT==2.6.0
python-dotenv==0.21.0
